Add structure of project
Some choices are made at this point about the shape of the project. Some
are obvious - we need c and python code to live here so we have suitable
src directory structures. Some are less obvious such as using maven for
the lifecycle. Realistically this should be just a better make but that
doesnt seem to exist right now. Would welcome suggestions.

Also scaffolded out the folder structure as well as adding placeholders
for various templates and tools. Not very useful but a reminder of how
I'd like things to work I suppose.